server/app/requirements.txt
pandas==2.0.0
numpy==1.21.5
flask==2.0.3
python_dotenv==0.14.0
apscheduler==3.7.0
statsmodels==0.13.2
yfinance==0.2.22
APScheduler==3.10.1